Use a deferred sync for handling the gravity sync
For the new reflux method, we need to use a deferred sync to handle
the correction to the potential from refluxing. The reason is that there
is a mismatch in time between grad_phi on different levels. We can correct
for this mismatch by scaling the delta in grad_phi by how far into the timestep
we are, solving for the sync potential, and only adding it to the full
gravitational potential for the purposes of calculating the boundary conditions
that are used on the level solves. At the end of the timestep on a level,
we then actually add it back into the full potential, so the deferral
never lasts more than the current coarse timestep.