Correct the deferred gravitational sync strategy
The correct strategy needs to update the potential temporarily before
recomputing the source terms on the coarse level. However, it cannot leave
the change there, as it defeats the purpose of the deferred sync (which
is to store the change separately, and add it back when the coarse grid
potential is linearly interpolated in time to create boundary conditions on
the fine grid -- storing it in the full state makes the time interpolation
wrong). So after we recompute the new-time source terms, we subtract off
the sync from the final time state (unless we've reached the end of the
timestep and thus the synchronization time between the coarse and fine
levels, in which case we leave it there as we no longer need to defer
the update).