Include the c-f grad(phi) mismatch in sync solve BCs
The mismatch in grad(phi) at a coarse-fine interface is a key part of
the gravitational synchronization solve (see Castro I, Equation 37).
Previously, it was left out of the sync solve. At the time this made some
sense because the boundary conditions were previously computed using only
the mass on the coarse grid, and only the contribution to the reflux
directly changes the mass on the coarse grid. The grad(phi) mismatch represents
a change to the coarse grid potential that must be made because of mass
motion on the fine grid, and it was arguably consistent to therefore
neglect it from the BC calculation at the time (consider that the point of
the sync solve is to get the same result that you would have gotten if
you had done a full multi-level composite solve instead, and the composite
solve BCs only included coarse grid data). Regardless, we now use data
from all levels in calculating the Poisson grid BCs (this surely is a better
idea for simulation accuracy) and consistency and accuracy therefore demand
that it should be included -- only with this contribution can the sync solve
be regarded as being equivalent to a full composite solve.